Make snake only moves when we wiggle
This is much better, can still tweak the radial addition distance and
snake length, but this is starting to feel like what I was looking
for.

A little weird that this is the first time that we've needed the state
for context when doing a sprite update function, maybe this should be
pushed back upstream? Maybe just the ability to get hold of the
`:held-keys` set?